# reduce ExtractIP allocations