```
benchmark                                                                                                                                   old ns/op     new ns/op     delta
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_54.55.101.102,_X-Forwarded-For:_nil-8                                     187           103           -44.92%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_54.55.101.102,_X-Forwarded-For:_54.56.103.104-8                           188           108           -42.55%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_54.56.103.104-8                                     301           218           -27.57%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_54.56.103.104_,_54.57.105.106_,_54.58.107.108-8     382           282           -26.18%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_nil-8                                               353           197           -44.19%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_,_X-Forwarded-For:_-8                                                     320           209           -34.69%
BenchmarkExtractIP/extractIP_remote:_,_X-Real-IP:_,_X-Forwarded-For:_-8                                                                     334           267           -20.06%

benchmark                                                                                                                                   old allocs     new allocs     delta
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_54.55.101.102,_X-Forwarded-For:_nil-8                                     4              1              -75.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_54.55.101.102,_X-Forwarded-For:_54.56.103.104-8                           4              1              -75.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_54.56.103.104-8                                     5              2              -60.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_54.56.103.104_,_54.57.105.106_,_54.58.107.108-8     5              2              -60.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_nil-8                                               5              2              -60.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_,_X-Forwarded-For:_-8                                                     5              2              -60.00%
BenchmarkExtractIP/extractIP_remote:_,_X-Real-IP:_,_X-Forwarded-For:_-8                                                                     6              3              -50.00%

benchmark                                                                                                                                   old bytes     new bytes     delta
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_54.55.101.102,_X-Forwarded-For:_nil-8                                     64            16            -75.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_54.55.101.102,_X-Forwarded-For:_54.56.103.104-8                           64            16            -75.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_54.56.103.104-8                                     80            32            -60.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_54.56.103.104_,_54.57.105.106_,_54.58.107.108-8     112           64            -42.86%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_nil-8                                               80            32            -60.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_,_X-Forwarded-For:_-8                                                     80            32            -60.00%
BenchmarkExtractIP/extractIP_remote:_,_X-Real-IP:_,_X-Forwarded-For:_-8                                                                     112           64            -42.86%
```