@axw 2018-04-10
Is ExtractIP really contributing much on the server?

I just went through the process of optimising similar code in the Go agent, you might like to reuse some of the code there. You can get this down to zero allocations in many cases.

```
benchcmp /tmp/old.txt /tmp/new.txt 
benchmark                                                                                                                                   old ns/op     new ns/op     delta
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_54.55.101.102,_X-Forwarded-For:_nil-4                                     118           33.3          -71.78%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_54.55.101.102,_X-Forwarded-For:_54.56.103.104-4                           144           32.6          -77.36%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_54.56.103.104-4                                     249           91.3          -63.33%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_54.56.103.104_,_54.57.105.106_,_54.58.107.108-4     306           107           -65.03%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_nil-4                                               209           81.7          -60.91%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_,_X-Forwarded-For:_-4                                                     225           97.4          -56.71%
BenchmarkExtractIP/extractIP_remote:_,_X-Real-IP:_,_X-Forwarded-For:_-4                                                                     252           73.7          -70.75%

benchmark                                                                                                                                   old allocs     new allocs     delta
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_54.55.101.102,_X-Forwarded-For:_nil-4                                     1              0              -100.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_54.55.101.102,_X-Forwarded-For:_54.56.103.104-4                           1              0              -100.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_54.56.103.104-4                                     2              0              -100.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_54.56.103.104_,_54.57.105.106_,_54.58.107.108-4     2              0              -100.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_nil-4                                               2              0              -100.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_,_X-Forwarded-For:_-4                                                     2              0              -100.00%
BenchmarkExtractIP/extractIP_remote:_,_X-Real-IP:_,_X-Forwarded-For:_-4                                                                     3              0              -100.00%

benchmark                                                                                                                                   old bytes     new bytes     delta
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_54.55.101.102,_X-Forwarded-For:_nil-4                                     16            0             -100.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_54.55.101.102,_X-Forwarded-For:_54.56.103.104-4                           16            0             -100.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_54.56.103.104-4                                     32            0             -100.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_54.56.103.104_,_54.57.105.106_,_54.58.107.108-4     64            0             -100.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_nil,_X-Forwarded-For:_nil-4                                               32            0             -100.00%
BenchmarkExtractIP/extractIP_remote:_10.11.12.13:8080,_X-Real-IP:_,_X-Forwarded-For:_-4                                                     32            0             -100.00%
BenchmarkExtractIP/extractIP_remote:_,_X-Real-IP:_,_X-Forwarded-For:_-4                                                                     64            0             -100.00%
```

```diff
diff --git a/utility/ip.go b/utility/ip.go
index 6c13a69..dbc1ddf 100644
--- a/utility/ip.go
+++ b/utility/ip.go
@@ -9,21 +9,30 @@ import (
 // Obtains the IP of a request, looking up X-Forwarded-For and X-Real-IP headers.
 // X-Forwarded-For has a list of IPs, of which the first is the one of the original client.
 // This value however might not be necessarily trusted, as it can be forged by a malicious user.
-func ExtractIP(r *http.Request) string {
-	realIP := r.Header.Get("X-Real-IP")
-	if realIP != "" {
+func ExtractIP(req *http.Request) string {
+	if realIP := req.Header.Get("X-Real-Ip"); realIP != "" {
 		return realIP
 	}
-
-	forwardedFor := r.Header.Get("X-Forwarded-For")
-	client := strings.Split(forwardedFor, ",")[0]
-	if client != "" {
-		return strings.TrimSpace(client)
+	if xff := req.Header.Get("X-Forwarded-For"); xff != "" {
+		if sep := strings.IndexRune(xff, ','); sep > 0 {
+			xff = xff[:sep]
+		}
+		return strings.TrimSpace(xff)
 	}
+	remoteAddr, _ := splitHost(req.RemoteAddr)
+	return remoteAddr
+}
 
-	ip, _, err := net.SplitHostPort(r.RemoteAddr)
+func splitHost(in string) (host, port string) {
+	if strings.LastIndexByte(in, ':') == -1 {
+		// In the common (relative to other "errors") case that
+		// there is no colon, we can avoid allocations by not
+		// calling SplitHostPort.
+		return in, ""
+	}
+	host, port, err := net.SplitHostPort(in)
 	if err != nil {
-		return r.RemoteAddr
+		return in, ""
 	}
-	return ip
+	return host, port
 }
```