@graphaelli 2018-04-10
Very cool, thanks. No it's not contributing much garbage even though it's called for most requests. This is still readable enough that it's a better update, thanks!